Improved Parameter Processing Performance
This commit dramatically improves the parameter parsing performance by separating out script parsing over the course of many ticks, rather than trying to parse
entire functions in a single tick.  Previously a large list of commands inside of a function would trigger the dreaded "Script Too Complex" issue.

This change effectively eliminates this issue by ensuring that only one command is processed per tick.  As such the script should parse any reasonably sized
function.  It is still technically possible to hit the Script Too Complex issue with a sufficiently large function, as all command parameters for a given function
are still parsed in a single tick.

Additionally, this change pulls the Command Parsing logic from Program into a separate ParsingEngine section to improve readability
