+++ EasyCommands/EasyCommands.csproj
@@ -144,6 +144,7 @@
     <Compile Include="BlockHandlers\WarheadBlockHandlers.cs" />
     <Compile Include="BlockHandlers\WheelBlockHandlers.cs" />
     <Compile Include="BlockHandlers\EngineBlockHandlers.cs" />
+    <Compile Include="CommandParsers\ParsingEngine.cs" />
     <Compile Include="CommandParsers\ParameterParsers.cs" />
     <Compile Include="CommandParsers\ParameterProcessors.cs" />
     <Compile Include="Commands\CommandParameters.cs" />
