+++ EasyCommands/Program.cs
@@ -195,122 +195,6 @@ namespace IngameScript {
             }
         }
 
-        bool ParseCommands() {
-            if (String.IsNullOrWhiteSpace(PROGRAM.Me.CustomData)) {
-                Info("Welcome to EasyCommands!");
-                Info("Add Commands to Custom Data");
-                return false;
-            } else if (!PROGRAM.Me.CustomData.Equals(customData)) {
-                customData = PROGRAM.Me.CustomData;
-                commandStrings = customData.Trim().Split(new[] { "\r\n", "\r", "\n" }, StringSplitOptions.None)
-                    .ToList();
-
-                Info("Parsing Custom Data");
-                functions.Clear();
-                ClearAllThreads();
-            }
-
-            if (commandStrings.Count == 0) return true;
-            Info("Parsing Commands.  Lines Left: " + commandStrings.Count);
-            ParseFunctions(commandStrings);
-
-            return commandStrings.Count == 0;
-        }
-
-        void ParseFunctions(List<String> commandStrings) {
-            var functionIndices = NewList<int>();
-            int implicitMainOffset = 1;
-            if (!commandStrings[0].StartsWith(":")) {
-                commandStrings.Insert(0, ":main");
-                implicitMainOffset--;
-            }
-
-            for (int i = commandStrings.Count - 1; i >= 0; i--) {
-                if (commandStrings[i].StartsWith(":")) { functionIndices.Add(i); }
-            }
-
-            //Parse Function Definitions
-            foreach (int i in functionIndices) {
-                String functionString = commandStrings[i].Remove(0, 1).Trim();
-                List<Token> nameAndParams = ParseTokens(functionString);
-                String functionName = nameAndParams[0].original;
-                nameAndParams.RemoveAt(0);
-                FunctionDefinition definition = new FunctionDefinition(functionName, nameAndParams.Select(t => t.original).ToList());
-                functions[functionName] = definition;
-            }
-
-            //Parse Function Commands and add to Definitions
-            int toParse = functionParseAmount;
-            foreach (int i in functionIndices) {
-                int startingLineNumber = i + 1 + implicitMainOffset;
-                String functionString = commandStrings[i].Remove(0, 1).Trim();
-                List<Token> nameAndParams = ParseTokens(functionString);
-                String functionName = nameAndParams[0].original;
-                Info("Parsing Function: " + functionName);
-                Command command = ParseCommand(commandStrings
-                    .GetRange(i + 1, commandStrings.Count - (i + 1))
-                    .Select(str => new CommandLine(str, startingLineNumber++))
-                    .Where(line => line.commandParameters.Count > 0)
-                    .ToList(), 0, true);
-                commandStrings.RemoveRange(i, commandStrings.Count - i);
-                if (!(command is MultiActionCommand)) command = new MultiActionCommand(NewList<Command>(command));
-                functions[functionName].function = (MultiActionCommand)command;
-                defaultFunction = functionName;
-                toParse--;
-                if (toParse == 0) break;//Exceeded # of Functions to parse for 1 tick
-            }
-        }
-
-        Command ParseCommand(List<CommandLine> commandStrings, int index, bool parseSiblings) {
-            var resolvedCommands = NewList<Command>();
-            while (index < commandStrings.Count - 1)//Parse Sibling & Child Commands, if any
-            {
-                CommandLine current = commandStrings[index];
-                CommandLine next = commandStrings[index + 1];
-                if (current.depth > next.depth) break; //End, break
-                if (current.depth < next.depth) { //I'm a parent of next line
-                    current.commandParameters.Add(new CommandReferenceParameter(ParseCommand(commandStrings, index + 1, true)));
-                    continue;
-                }
-                if (next.commandParameters.Count > 0 && next.commandParameters[0] is ElseCommandParameter) {//Handle Otherwise
-                    current.commandParameters.Add(next.commandParameters[0]);
-                    next.commandParameters.RemoveAt(0);
-                    current.commandParameters.Add(new CommandReferenceParameter(ParseCommand(commandStrings, index + 1, false)));
-                    continue;
-                }
-
-                if (!parseSiblings) break; //Only parsing myself
-                resolvedCommands.Add(ParseCommand(current.commandParameters, current.lineNumber));
-                commandStrings.RemoveAt(index);
-            }
-
-            //Parse Last one, which has become current
-            resolvedCommands.Add(ParseCommand(commandStrings[index].commandParameters, commandStrings[index].lineNumber));
-            commandStrings.RemoveAt(index);
-
-            return resolvedCommands.Count > 1 ? new MultiActionCommand(resolvedCommands) : resolvedCommands[0];
-        }
-
-        public Command ParseCommand(String commandLine, int lineNumber = 0) =>
-            ParseCommand(ParseCommandParameters(ParseTokens(commandLine)), lineNumber);
-
-        Command ParseCommand(List<CommandParameter> parameters, int lineNumber) {
-            CommandReferenceParameter command = ParseParameters<CommandReferenceParameter>(parameters);
-
-            if (command == null) throw new Exception("Unable to parse command from command parameters at line number: " + lineNumber);
-            return command.value;
-        }
-        class CommandLine {
-            public int depth, lineNumber;
-            public List<CommandParameter> commandParameters;
-
-            public CommandLine(String command, int line) {
-                depth = command.TakeWhile(Char.IsWhiteSpace).Count();
-                commandParameters = PROGRAM.ParseCommandParameters(PROGRAM.ParseTokens(command));
-                lineNumber = line;
-            }
-        }
-
         public class Thread {
             public Command Command { get; set; }
             public Dictionary<String, Variable> threadVariables = NewDictionary<string, Variable>();
